refactor(rules): isolate pausal mark rules
